Decks contains informations about limits and value
This means that the tree structure can easily be used in the deck
directly instead of creating an add-hoc method.
It also means that all computations for limits, values, and possibly
future values, can also be computed directly in the deck object. It
leads to a clearer vue of the meaning of each object.