Allow to remove useless column from browser
Currently, if for some reason, the browser have an active columns
which does not exists, instead of showing any information to the user
related to the column, it shows the name of the alphabetically first
column. This makes little sens. And worse, this column can't be
removed. Indeed, intuitively the user may want to click on the name
shown in the column to remove it. Instead it'll add the real first
column (or remove it if it really was here. In which case it only
removes the column with actual information.)

I thus do believe that it is slighly better to allow the user to
remove any useless columns (and not to add them back. At worst, the
add-on creating those columns will deal with them)
To do so, I add to show the "type" of the columns, since I don't have
access to their name.

Another solution which would makes sens too would simply be to hide
those columns from the users. Note that this would create potentially
a problem because in this case they may have 0 columns shown if all
active columns are columns anki does not know