remove decks.update
This was used only twice. Once was useless, since it's related to
exporting. And there is no need to do so much thing in an exported
collection.

The other was while syncing. And I think that just reloading the set
of decks later is a better use of the time